+++ connection_util.go
@@ -149,7 +149,7 @@ func (sc *snowflakeConn) processFileTransfer(
 	if err != nil {
 		return nil, err
 	}
-	if sfa.options.GetFileToStream {
+	if sfa.options != nil && sfa.options.GetFileToStream {
 		if err := writeFileStream(ctx, sfa.streamBuffer); err != nil {
 			return nil, err
 		}